move kpi_data to mixed_precision common package
