+++ model_compression_toolkit/keras/kpi_data_facade.py
@@ -16,7 +16,7 @@
 from typing import Callable
 
 from model_compression_toolkit.common.hardware_representation import FrameworkHardwareModel
-from model_compression_toolkit.common.kpi_data import compute_kpi_data
+from model_compression_toolkit.common.mixed_precision.kpi_data import compute_kpi_data
 from model_compression_toolkit.common.framework_info import FrameworkInfo
 from model_compression_toolkit.common.quantization.quantization_config import QuantizationConfig
 from model_compression_toolkit.common.quantization.quantization_config import DEFAULTCONFIG
@@ -30,7 +30,6 @@ if importlib.util.find_spec("tensorflow") is not None\
     from model_compression_toolkit.keras.default_framework_info import DEFAULT_KERAS_INFO
     from model_compression_toolkit.keras.keras_implementation import KerasImplementation
     from tensorflow.keras.models import Model
-
 
 
     def keras_kpi_data(in_model: Model,
